feature/refactor FileSystemList
