+++ apps/ui/src/core/file-system/components/FileSystemList/types.ts
@@ -1,0 +1,17 @@
+import { FileTree } from "@/types";
+
+export type GitStatus = {
+  working_dir?: string;
+  index?: string;
+};
+
+export type ExtendedFileTree = FileTree & {
+  id: string;
+  parent?: string;
+  isTemporary?: boolean;
+  fileKind?: "file" | "void" | undefined;
+  children?: ExtendedFileTree[];
+  lazy?: boolean;
+  git?: GitStatus;
+  aggregatedGitStatus?: GitStatus;
+};
